Remove defunct code
Remove the deprecation warning on `Controller#Error` access. Actually deprecate it.

Remove the helpers lib. Most of the code wasn't being used. Moved the small parts that were being used into the form controller so they can be extended in implementations.

Removed the two-tier error class. Since the next step is to move message parsing into separate, render-time code then there is no need to make that distinction.
