+++ index.js
@@ -10,10 +10,8 @@ class FormController extends mix(Controller).with(Session, Hooks) {}
 FormController.validators = require('./lib/validation/validators');
 FormController.formatters = require('./lib/formatting/formatters');
 
-FormController.ValidationError = require('./lib/error');
-FormController.Error = mix(FormController.ValidationError).with(require('./lib/deprecate-error'));
+FormController.ValidationError = require('./lib/validation-error');
 
 module.exports = FormController;
 module.exports.Controller = Controller;
 module.exports.BaseController = require('./lib/base-controller');
-module.exports.BaseError = require('./lib/base-error');
