+++ lib/base-controller.js
@@ -7,9 +7,7 @@ const EventEmitter = require('events').EventEmitter;
 const debug = require('debug')('hmpo:form');
 const dataFormatter = require('./formatting');
 const dataValidator = require('./validation');
-const ErrorClass = require('./base-error');
-const Deprecate = require('./deprecate-error');
-const mix = require('mixwith').mix;
+const ErrorClass = require('./validation-error');
 
 module.exports = class BaseController extends EventEmitter {
   constructor(options) {
@@ -23,9 +21,7 @@ module.exports = class BaseController extends EventEmitter {
     options.defaultFormatters = options.defaultFormatters || ['trim', 'singlespaces', 'hyphens'];
     options.fields = options.fields || {};
     this.options = options;
-
     this.ValidationError = ErrorClass;
-    this.Error = mix(ErrorClass).with(Deprecate);
 
     this.router = express.Router({ mergeParams: true });
   }
@@ -137,7 +133,7 @@ module.exports = class BaseController extends EventEmitter {
       if (error) {
         const errorKey = error.group || error.key;
         return Object.assign({}, errs, {
-          [errorKey]: new this.ValidationError(errorKey, error, req, res)
+          [errorKey]: new this.ValidationError(errorKey, error)
         });
       }
       return errs;
