+++ lib/controller.js
@@ -3,11 +3,7 @@
 const _ = require('lodash');
 const i18nLookup = require('i18n-lookup');
 const Mustache = require('mustache');
-const helpers = require('./util/helpers');
 const BaseController = require('./base-controller');
-const ErrorClass = require('./error');
-const Deprecate = require('./deprecate-error');
-const mix = require('mixwith').mix;
 
 const omitField = (field, req) => field.useWhen && (typeof field.useWhen === 'string'
   ? req.sessionModel.get(field.useWhen) !== 'true'
@@ -16,8 +12,6 @@ const omitField = (field, req) => field.useWhen && (typeof field.useWhen === 'st
 module.exports = class Controller extends BaseController {
   constructor(options) {
     super(options);
-    this.ValidationError = ErrorClass;
-    this.Error = mix(ErrorClass).with(Deprecate);
     this.confirmStep = options.confirmStep || '/confirm';
   }
 
@@ -120,12 +114,30 @@ module.exports = class Controller extends BaseController {
       fields,
       route,
       baseUrl: req.baseUrl,
-      title: helpers.getTitle(route, lookup, req.form.options.fields, res.locals),
-      intro: helpers.getIntro(route, lookup, res.locals),
+      title: this.getTitle(route, lookup, req.form.options.fields, res.locals),
+      intro: this.getIntro(route, lookup, res.locals),
       backLink: this.getBackLink(req, res),
       nextPage: this.getNextStep(req, res),
       errorLength: this.getErrorLength(req, res)
     }, stepLocals);
+  }
+
+  getTitle(route, lookup, fields, locals) {
+    let fieldName = '';
+    if (_.size(fields)) {
+      fieldName = Object.keys(fields)[0];
+    }
+    return lookup([
+      `pages.${route}.header`,
+      `fields.${fieldName}.label`,
+      `fields.${fieldName}.legend`
+   ], locals);
+  }
+
+  getIntro(route, lookup, locals) {
+    return lookup([
+      `pages.${route}.intro`
+    ], locals);
   }
 
   getErrorLength(req, res) {
